Store Telegram IDs as bigint

+++ app/models.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from enum import StrEnum
 import re
 
-from sqlalchemy import DateTime, Integer, String, Text, func
+from sqlalchemy import BigInteger, DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -30,9 +30,10 @@ class Order(Base):
     comment: Mapped[str] = mapped_column(Text, nullable=False, default="")
     status: Mapped[str] = mapped_column(String(64), nullable=False, default=OrderStatus.NEW.value)
     forwarded_from: Mapped[str | None] = mapped_column(String(255), nullable=True)
-    created_by_telegram_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # Telegram IDs can be greater than the maximum value of a 32-bit INTEGER.
+    created_by_telegram_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     created_by_name: Mapped[str | None] = mapped_column(String(255), nullable=True)
-    updated_by_telegram_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    updated_by_telegram_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     updated_by_name: Mapped[str | None] = mapped_column(String(255), nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     updated_at: Mapped[datetime] = mapped_column(
